+++ src/IsCollectionType.php
@@ -13,6 +13,8 @@ use FireMidge\ValueObject\Exception\ValueNotFound;
  * e.g. a list of e-mail addresses.
  *
  * You can implement \Iterator in your class.
+ *
+ * @template T of mixed
  */
 trait IsCollectionType
 {
@@ -36,6 +38,8 @@ trait IsCollectionType
 
     /**
      * Creates a new instance from an array of values.
+     *
+     * @param T[] $values
      */
     public static function fromArray(array $values) : static
     {
@@ -53,10 +57,12 @@ trait IsCollectionType
     /**
      * Returns a new instance with $addedValue added to the list.
      *
+     * @param T $addedValue
+     *
      * @throws InvalidValue  If values must be unique and $addedValue is a duplicate.
      * @throws InvalidValue  If other validation checks have been set up and $addedValue is invalid (e.g. invalid type).
      */
-    public function withValue($addedValue) : static
+    public function withValue(mixed $addedValue) : static
     {
         $addedValue = $this->transformEach($addedValue);
 
@@ -71,6 +77,8 @@ trait IsCollectionType
 
     /**
      * Returns a new instance with all of $addedValues added to the list.
+     *
+     * @param T[] $addedValues
      *
      * @throws InvalidValue  If values must be unique and any of $addedValues is a duplicate.
      * @throws InvalidValue  If other validation checks have been set up and any of $addedValues is invalid (e.g. invalid type).
@@ -89,6 +97,8 @@ trait IsCollectionType
      * Returns a new instance without the value if the value previously existed.
      * If the value did not exist, it will return a new instance with the same
      * values.
+     *
+     * @param T $value
      *
      * @throws InvalidValue  If validation checks have been set up and $value is invalid.
      */
@@ -110,6 +120,8 @@ trait IsCollectionType
      * Returns a new instance without any of the values in $values.
      * Any that did not previously exist are ignored.
      *
+     * @param T[] $values
+     *
      * @throws InvalidValue  If validation checks have been set up and any of $values is invalid.
      */
     public function withoutValues(array $values) : static
@@ -122,15 +134,16 @@ trait IsCollectionType
         return $instance;
     }
 
-
     /**
      * Same as `withoutValue` but throws an exception when trying to
      * remove a value that did not exist.
      *
+     * @param T $value
+     *
      * @throws ValueNotFound If the value did not previously exist in the list.
      * @throws InvalidValue  If validation checks have been set up and $value is invalid.
      */
-    public function tryWithoutValue($value) : static
+    public function tryWithoutValue(mixed $value) : static
     {
         $this->validateEach($value);
         $newValues = $this->cloneValues($this->values);
@@ -145,6 +158,8 @@ trait IsCollectionType
     /**
      * Same as `withoutValues` but throws an exception when trying to
      * remove a value that did not exist.
+     *
+     * @param T[] $values
      *
      * @throws ValueNotFound If any of the values did not previously exist in the list.
      * @throws InvalidValue  If validation checks have been set up and any of $values is invalid.
@@ -162,11 +177,13 @@ trait IsCollectionType
     /**
      * Returns true if this list contains $value.
      *
+     * @param T $value
+     *
      * Note that checks are not performed with strict types, but if you set up a
      * type check in validateEach(), it will be performed before making the comparison.
      * You can also override this method to explicitly perform a strict comparison.
      */
-    public function contains($value) : bool
+    public function contains(mixed $value) : bool
     {
         $this->validateEach($value);
         return (in_array($value, $this->values));
@@ -181,7 +198,7 @@ trait IsCollectionType
      *                                 E.g. fn($v) => $v.name === 'Susan'
      *                                 or: fn($v, $k) => $k === 10
      *
-     * @return mixed  The element matching the criteria, or `null` if no match was found.
+     * @return ?T  The element matching the criteria, or `null` if no match was found.
      */
     public function find(callable $searchCallback) : mixed
     {
@@ -216,17 +233,142 @@ trait IsCollectionType
         return null;
     }
 
+    /**
+     * Merges $other into itself.
+     * This modifies the current instance.
+     *
+     * @param object|T[]|array $other  Can be an instance of the same class, an instance of a different class
+     *                                 holding the same types of value, or an array.
+     */
+    public function merge(object|array $other) : void
+    {
+        $new = $this->withValues($this->convertToArray($other));
+        $this->values = $new->values;
+    }
+
+    /**
+     * Returns a new instance which contains the values of both this instance and $other.
+     *
+     * @param object|T[]|mixed[] $other Can be an instance of the same class, an instance of
+     *                                  a different class holding the same types of value, or an array.
+     */
+    public function withMerged(object|array $other) : static
+    {
+        return $this->withValues($this->convertToArray($other));
+    }
+
+    /**
+     * Removes the last element from itself and returns it.
+     * This modifies the current instance.
+     *
+     * @return ?T
+     */
+    public function pop() : mixed
+    {
+        return array_pop($this->values);
+    }
+
+    /**
+     * Similar to split(), with 2 key differences:
+     * 1) Instead of specifying the number of elements to keep, this lets
+     *    you specify the number of elements to remove from the end of the array.
+     * 2) The elements are returned in reverse order.
+     *    Removing the last 2 elements from 'a', 'b', 'c' results in 'c', 'b'.
+     *
+     * This modifies the current instance.
+     *
+     * @param int $numberOfElements How many elements to remove from the end of the array.
+     *                              The removed elements will be returned in a new instance
+     *                              in reverse order.
+     */
+    public function popMultiple(int $numberOfElements) : static
+    {
+        $poppedCards = static::empty();
+        for ($i = 0; $i < $numberOfElements; $i++) {
+            $poppedElement = $this->pop();
+
+            if ($poppedElement === null) {
+                break;
+            }
+
+            $poppedCards = $poppedCards->withValue($poppedElement);
+        }
+
+        return $poppedCards;
+    }
+
+    /**
+     * Similar to popMultiple(), with 2 key differences:
+     *  1) Instead of specifying the number of elements to remove, this lets
+     *     you specify the number of elements to keep in the original array.
+     *     Note that you can force this by using a negative $numberOfValuesToKeep.
+     *  2) The elements are returned in the same order they were in the original array.
+     *     Removing the last 2 elements from 'a', 'b', 'c' results in 'b', 'c'.
+     *
+     * This modifies the current instance.
+     *
+     * @param int $numberOfValuesToKeep  How many elements should remain in the current instance.
+     *                                   The removed elements are returned in a new instance.
+     *                                   If this is a negative number, it changes to the number
+     *                                   of elements to remove.
+     */
+    public function split(int $numberOfValuesToKeep) : static
+    {
+        $numberOfValuesToRemove = $numberOfValuesToKeep < 0
+            ? abs($numberOfValuesToKeep)
+            : count($this->values) - $numberOfValuesToKeep;
+
+        if ($numberOfValuesToRemove <= 0) {
+            return static::empty();
+        }
+
+        return $this->popMultiple($numberOfValuesToRemove)->withReversedOrder();
+    }
+
+    /**
+     * Returns a new instance, where all the values are in the reversed order.
+     *
+     * If the current instance has ['a', 'b', 'c'], the new instance will
+     * have ['c', 'b', 'a'].
+     */
+    public function withReversedOrder() : static
+    {
+        $values = $this->values;
+        return static::fromArray(array_reverse($values));
+    }
+
+    /**
+     * Returns the first element contained within itself.
+     * This moves the pointer of the current instance but does not remove any elements.
+     *
+     * @return ?T
+     */
     public function first() : mixed
     {
         reset($this->values);
-        return current($this->values);
+        return $this->current();
     }
 
+    /**
+     * Returns the first element contained within itself.
+     * This moves the pointer of the current instance but does not remove any elements.
+     *
+     * @return ?T
+     */
     public function last() : mixed
     {
-        return end($this->values);
+        $lastElement = end($this->values);
+
+        if ($lastElement === false) {
+            return null;
+        }
+
+        return $lastElement;
     }
 
+    /**
+     * @return ?T
+     */
     public function current() : mixed
     {
         $item = current($this->values);
@@ -238,26 +380,15 @@ trait IsCollectionType
         return $item;
     }
 
-    public function next() : mixed
+
+    public function next() : void
     {
-        $item = next($this->values);
-
-        if ($item === false) {
-            return null;
-        }
-
-        return $item;
+        next($this->values);
     }
 
-    public function previous() : mixed
+    public function previous() : void
     {
-        $item = prev($this->values);
-
-        if ($item === false) {
-            return null;
-        }
-
-        return $item;
+        prev($this->values);
     }
 
     public function key() : mixed
@@ -275,8 +406,15 @@ trait IsCollectionType
         reset($this->values);
     }
 
+    public function shuffle() : void
+    {
+        shuffle($this->values);
+    }
+
     /**
      * Converts this list back into a primitive array.
+     *
+     * @return T[]
      */
     public function toArray() : array
     {
@@ -392,9 +530,9 @@ trait IsCollectionType
      *
      * @param mixed  $value  The input value to transform.
      *
-     * @return mixed
+     * @return T
      */
-    protected function transformEach($value)
+    protected function transformEach(mixed $value) : mixed
     {
         return $value;
     }
@@ -402,7 +540,7 @@ trait IsCollectionType
     /**
      * Override this to provide custom handling of duplicate values within an array.
      *
-     * @param mixed $values The list of values to be added to the instance, which contain duplicates.
+     * @param T[] $values The list of values to be added to the instance, which contain duplicates.
      *
      * @throws DuplicateValue If there is/are duplicate value(s) that cannot be accepted.
      */
@@ -427,7 +565,7 @@ trait IsCollectionType
      * Override this to provide custom handling of adding a value that is a duplicate of an
      * already existing value.
      *
-     * @param mixed $value The new value to be added, which is a duplicate of one of the
+     * @param T $value The new value to be added, which is a duplicate of one of the
      *                     pre-existing values.
      *
      * @return array The full array of values to be stored against the object. You will have handled
@@ -460,11 +598,7 @@ trait IsCollectionType
 
     private function isEqualToObject(object $other) : bool
     {
-        if (method_exists($other, 'toArray')) {
-            return $this->isEqualToArray($other->toArray());
-        }
-
-        return $this->isEqualToArray((array) $other);
+        return $this->isEqualToArray($this->convertToArray($other));
     }
 
     private function cloneValues(array $values) : array
@@ -483,7 +617,7 @@ trait IsCollectionType
      * but `validateEach` is always called before calling
      * this method.
      */
-    private function getIndexForValue($value) : int
+    private function getIndexForValue(mixed $value) : int
     {
         // Not doing a strict comparison here by default,
         // as this would not work as intended for objects.
@@ -494,5 +628,18 @@ trait IsCollectionType
         }
 
         return $index;
+    }
+
+    private function convertToArray(object|array $other) : array
+    {
+        if (is_array($other)) {
+            return $other;
+        }
+
+        if (method_exists($other, 'toArray')) {
+            return $other->toArray();
+        }
+
+        return (array) $other;
     }
 }